requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.4.26
frozenlist==1.6.0
greenlet==3.2.2
idna==3.10
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.20.0
